+++ src/test/java/com/contoso/liveservicetests/TestCore.java
@@ -7,15 +7,11 @@
 package com.contoso.liveservicetests;
 
 import com.contoso.helpers.HelperUtils;
-import com.microsoft.azure.datalake.store.ADLException;
-import com.microsoft.azure.datalake.store.ADLStoreClient;
-import com.microsoft.azure.datalake.store.SyncFlag;
+import com.microsoft.azure.datalake.store.*;
 import com.microsoft.azure.datalake.store.oauth2.AzureADAuthenticator;
 import com.microsoft.azure.datalake.store.oauth2.AzureADToken;
-import com.microsoft.azure.datalake.store.Core;
-import com.microsoft.azure.datalake.store.OperationResponse;
-import com.microsoft.azure.datalake.store.RequestOptions;
 
+import com.microsoft.azure.datalake.store.retrypolicies.NoRetryPolicy;
 import org.junit.AfterClass;
 import org.junit.Assume;
 import org.junit.BeforeClass;
@@ -23,6 +19,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.*;
 
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
@@ -66,6 +63,7 @@ public class TestCore {
     public void createSmallFileWithOverWrite() throws IOException {
         Assume.assumeTrue(testsEnabled);
         String filename = directory + "/" + "Core.CreateSmallFileWithOverWrite.txt";
+        System.out.println("Running createSmallFileWithOverWrite");
 
         byte [] contents = HelperUtils.getSampleText1();
         putFileContents(filename, contents, true);
@@ -79,6 +77,7 @@ public class TestCore {
     public void createSmallFileWithNoOverwrite() throws IOException {
         Assume.assumeTrue(testsEnabled);
         String filename = directory + "/" + "Core.CreateSmallFileWithNoOverwrite.txt";
+        System.out.println("Running createSmallFileWithNoOverwrite");
 
         byte [] contents = HelperUtils.getSampleText1();
         putFileContents(filename, contents, false);
@@ -88,10 +87,155 @@ public class TestCore {
         assertTrue("file contents should match", Arrays.equals(b, contents));
     }
 
+
+    @Test
+    public void createEmptyFileWithConcurrentAppend() throws IOException {
+        Assume.assumeTrue(false);  // pending change to server behavior
+        String filename = directory + "/" + "Core.createEmptyFileWithConcurrentAppend.txt";
+        System.out.println("Running createEmptyFileWithConcurrentAppend");
+
+        byte [] contents = null;
+        RequestOptions opts = new RequestOptions();
+        opts.retryPolicy = new NoRetryPolicy();
+        OperationResponse resp = new OperationResponse();
+        Core.concurrentAppend(filename, contents, 0, 0, true,
+                client, opts, resp);
+        if (!resp.successful) {
+            throw client.getExceptionFromResponse(resp, "Error in ConcurrentAppend with null content " + filename);
+        }
+
+        DirectoryEntry de = dir(filename);
+        assertTrue("File type should be FILE", de.type == DirectoryEntryType.FILE);
+        assertTrue("File length in DirectoryEntry should be 0 for null-content file", de.length == 0);
+
+        byte[] b = getFileContents(filename, contents.length * 2);
+        assertTrue("file length should be 0 for null content", b.length == 0);
+    }
+
+
+    @Test
+    public void create0LengthFileWithConcurrentAppend() throws IOException {
+        Assume.assumeTrue(false);  // pending change to server behavior
+        String filename = directory + "/" + "Core.create0LengthFileWithConcurrentAppend.txt";
+        System.out.println("Running create0LengthFileWithConcurrentAppend");
+
+        byte [] contents = new byte[0];
+        RequestOptions opts = new RequestOptions();
+        opts.retryPolicy = new NoRetryPolicy();
+        OperationResponse resp = new OperationResponse();
+        Core.concurrentAppend(filename, contents, 0, 0, true,
+                client, opts, resp);
+        if (!resp.successful) {
+            throw client.getExceptionFromResponse(resp, "Error in ConcurrentAppend with 0-len content " + filename);
+        }
+
+        DirectoryEntry de = dir(filename);
+        assertTrue("File type should be FILE for 0-len file", de.type == DirectoryEntryType.FILE);
+        assertTrue("File length in DirectoryEntry should be 0 for 0-len file", de.length == 0);
+
+        byte[] b = getFileContents(filename, contents.length * 2);
+        assertTrue("file length should be 0 for 0-len file", b.length == 0);
+    }
+
+
+    @Test
+    public void createSmallFileWithConcurrentAppend() throws IOException {
+        Assume.assumeTrue(testsEnabled);
+        String filename = directory + "/" + "Core.createSmallFileWithConcurrentAppend.txt";
+        System.out.println("Running createSmallFileWithConcurrentAppend");
+
+        byte [] contents = HelperUtils.getSampleText1();
+        RequestOptions opts = new RequestOptions();
+        opts.retryPolicy = new NoRetryPolicy();
+        OperationResponse resp = new OperationResponse();
+        Core.concurrentAppend(filename, contents, 0, contents.length, true,
+                client, opts, resp);
+        if (!resp.successful) {
+            throw client.getExceptionFromResponse(resp, "Error in ConcurrentAppend with small content " + filename);
+        }
+
+        getFileContents(filename, contents.length * 2); // read, to force metadata sync
+
+        DirectoryEntry de = dir(filename);
+        assertTrue("File type should be FILE", de.type == DirectoryEntryType.FILE);
+        assertTrue("File length in DirectoryEntry should match (" + de.length + "!=" + contents.length + ")",
+                de.length == contents.length);
+
+        byte[] b = getFileContents(filename, contents.length * 2);
+        assertTrue("file length should match after ConcurrentAppend", b.length == contents.length);
+        assertTrue("file contents should match after ConcurrentAppend", Arrays.equals(contents, b));
+    }
+
+    @Test(expected = ADLException.class)
+    public void concurrentAppendWithoutAutocreate() throws IOException {
+        Assume.assumeTrue(testsEnabled);
+        String filename = directory + "/" + "Core.concurrentAppendWithoutAutocreate.txt";
+        System.out.println("Running concurrentAppendWithoutAutocreate");
+
+        byte [] contents = HelperUtils.getSampleText1();
+        RequestOptions opts = new RequestOptions();
+        opts.retryPolicy = new NoRetryPolicy();
+        OperationResponse resp = new OperationResponse();
+        Core.concurrentAppend(filename, contents, 0, contents.length, false,
+                client, opts, resp);
+        if (!resp.successful) {
+            throw client.getExceptionFromResponse(resp, "(expected) Exception from concurrentAppend " + filename);
+        }
+        // should throw exception
+    }
+
+
+    @Test
+    public void concurrentAppendToExistingFile() throws IOException {
+        Assume.assumeTrue(testsEnabled);
+        String filename = directory + "/" + "Core.concurrentAppendToExistingFile.txt";
+        System.out.println("Running concurrentAppendToExistingFile");
+
+        ByteArrayOutputStream bos = new ByteArrayOutputStream(16 * 1024);
+
+        byte [] contents = HelperUtils.getSampleText1();
+        RequestOptions opts = new RequestOptions();
+        opts.retryPolicy = new NoRetryPolicy();
+        OperationResponse resp = new OperationResponse();
+        Core.concurrentAppend(filename, contents, 0, contents.length, true,
+                client, opts, resp);
+        if (!resp.successful) {
+            throw client.getExceptionFromResponse(resp, "Exception from concurrentAppend " + filename);
+        }
+        bos.write(contents);
+
+        contents = HelperUtils.getSampleText2();
+        opts = new RequestOptions();
+        opts.retryPolicy = new NoRetryPolicy();
+        resp = new OperationResponse();
+        Core.concurrentAppend(filename, contents, 0, contents.length, false,
+                client, opts, resp);
+        if (!resp.successful) {
+            throw client.getExceptionFromResponse(resp, "Exception from concurrentAppend " + filename);
+        }
+        bos.write(contents);
+
+        bos.close();
+        byte[] b1 = bos.toByteArray();
+
+        getFileContents(filename, contents.length * 2); // read, to force metadata sync
+
+        DirectoryEntry de = dir(filename);
+        assertTrue("File type should be FILE", de.type == DirectoryEntryType.FILE);
+        assertTrue("File length in DirectoryEntry should match", de.length == b1.length);
+
+        byte[] b = getFileContents(filename, b1.length * 2);
+        assertTrue("file length should match after ConcurrentAppend", b.length == b1.length);
+        assertTrue("file contents should match after ConcurrentAppend", Arrays.equals(b1, b));
+    }
+
+
+
     @Test
     public void create4MBFile() throws IOException {
-        Assume.assumeTrue(testsEnabled);
+        Assume.assumeTrue(false);  // subsumed by TestFileSdk tests
         String filename = directory + "/" + "Core.Create4MBFile.txt";
+        System.out.println("Running create4MBFile");
 
         byte [] contents = HelperUtils.getRandomBuffer(4 * 1024 * 1024);
         putFileContents(filename, contents, true);
@@ -103,8 +247,9 @@ public class TestCore {
 
     @Test
     public void create5MBFile() throws IOException {
-        Assume.assumeTrue(testsEnabled);
+        Assume.assumeTrue(false);  // subsumed by TestFileSdk tests
         String filename = directory + "/" + "Core.Create5MBFile.txt";
+        System.out.println("Running create5MBFile");
 
         byte [] contents = HelperUtils.getRandomBuffer(11 * 1024 * 1024);
         putFileContents(filename, contents, true);
@@ -118,6 +263,7 @@ public class TestCore {
     public void createOverwriteFile() throws IOException {
         Assume.assumeTrue(testsEnabled);
         String filename = directory + "/" + "Core.CreateOverWriteFile.txt";
+        System.out.println("Running createOverwriteFile");
 
         byte[] contents = HelperUtils.getSampleText1();
         putFileContents(filename, contents, true);
@@ -134,6 +280,7 @@ public class TestCore {
     public void createNoOverwriteFile() throws IOException {
         Assume.assumeTrue(testsEnabled);
         String filename = directory + "/" + "Core.CreateNoOverWriteFile.txt";
+        System.out.println("Running createNoOverwriteFile");
 
         byte[] contents = HelperUtils.getSampleText1();
         putFileContents(filename, contents, true);
@@ -161,16 +308,14 @@ public class TestCore {
             RequestOptions opts = new RequestOptions();
             OperationResponse resp = new OperationResponse();
             InputStream in = Core.open(filename, count, 0, null, client, opts, resp);
-            System.out.format("Open completed. Current count=%d%n", count);
-            if (!resp.successful) throw client.getExceptionFromResponse(resp, "Error reading from file " + filename);
             if (resp.httpResponseCode == 403 || resp.httpResponseCode == 416) {
                 eof = true;
                 continue;
             }
+            if (!resp.successful) throw client.getExceptionFromResponse(resp, "Error reading from file " + filename);
             int bytesRead;
             while ((bytesRead = in.read(b, count, b.length - count)) != -1) {
                 count += bytesRead;
-                System.out.format("    read: %d, cumulative:%d%n", bytesRead, count);
                 if (count >= b.length) break;
             }
             in.close();
@@ -178,4 +323,16 @@ public class TestCore {
         byte[] b2 = Arrays.copyOfRange(b, 0, count);
         return b2;
     }
+
+    private DirectoryEntry dir(String filename) throws IOException {
+        RequestOptions opts = new RequestOptions();
+        opts.retryPolicy = new NoRetryPolicy();
+        OperationResponse resp = new OperationResponse();
+        DirectoryEntry de = Core.getFileStatus(filename, UserGroupRepresentation.OID, client, opts, resp);
+        if (!resp.successful) {
+            throw client.getExceptionFromResponse(resp, "Error in ConcurrentAppend with null content " + filename);
+        }
+        return de;
+    }
+
 }